6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+
 /**
  * main - print the string in the put function
  *
@@ -7,6 +8,7 @@
  * this program prints "Programming is like building a multilingual puzzle
  * Return: 0
  */
+
 int main(void)
 {
 	char c;
@@ -15,11 +17,11 @@ int main(void)
 	long long lli;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", sizeof(i));
-	printf("size of a long int: %lu byte(s)\n", sizeof(li));
-	printf("size of a long long int: %lu byte(s)\n", sizeof(lli);
-	printf("size of a float: %lu byte(s)\n", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(li));
+	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(lli);
+	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 	}
